feat: remove global.ROARR.prepend
This is technically a breaking change – if you had global.ROARR.prepend configured, it will no longer be appended to your logs. I have never had practical use case to this. Turns out that combination of instance ID and sequence ID is enough to create a logical timeline of the logs.